Extract the zsh config (#73)
* Updates nix flake and home manager config

Updates the nix flake to include nixd and nixvim directly.
Adds nixd to the home manager packages.
Adds .net sdk's to the home manager packages.
Updates dprint config to include yaml formatting.

* Refactors ZSH configuration into a separate file

Moves the ZSH configuration from the main home.nix file into a dedicated module for better organization and maintainability.

This change promotes modularity and improves the readability of the home configuration.

* Updates flake.lock dependencies

Updates the flake.lock file to reflect the latest
versions of dependencies, ensuring compatibility
and access to the newest features and bug fixes.